# Limit number of threads used by heavyweight processes for photo scanning