@derneuere 2021-11-28
Tested it on my system and everything works as expected. Code also looks good to me 👍 
Thanks for the pull request!